+++ src/05-seed-location.h
@@ -5,5 +5,5 @@ using namespace std;
 class Solution_05
 {
     public:
-    int part1(vector<int> seeds, vector<vector<vector<int>>> mappings);
+    int part1(vector<long> seeds, vector<vector<vector<long>>> mappings);
 };
